@@ -30,18 +30,16 @@ interface AuthorizableInterface extends IDrewlabsAuthorizable
     public function getAuthorizationGroups();
 
     /**
-     * Set the list of authorizations on the AuthorizableInterface instance
-     * 
-     * @param array $value 
-     * @return self 
+     * Set the list of authorizations on the AuthorizableInterface instance.
+     *
+     * @return self
      */
     public function setAuthorizations(array $value = []);
 
     /**
-     * Set the list of authorizations groups on the AuthorizableInterface instance
-     * 
-     * @param array $value 
-     * @return self 
+     * Set the list of authorizations groups on the AuthorizableInterface instance.
+     *
+     * @return self
      */
     public function setAuthorizationGroups(array $value = []);
 }
