@@ -1,18 +1,27 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file is part of the Drewlabs package.
+ *
+ * (c) Sidoine Azandrew <azandrewdevelopper@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace Drewlabs\Contracts\Support\Actions;
 
-use Closure;
 interface ActionHandler
 {
     /**
      * Handle a specific action and return a result. The callback is provided for handling
-     * action after the action is performed
-     * 
-     * @param Action $action
-     * @param Closure $callback
-     * 
+     * action after the action is performed.
+     *
+     * @param \Closure $callback
+     *
      * @return ActionResult
      */
-    public function handle(Action $action, Closure $callback = null);
+    public function handle(Action $action, ?\Closure $callback = null);
 }
