@@ -1,5 +1,16 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file is part of the Drewlabs package.
+ *
+ * (c) Sidoine Azandrew <azandrewdevelopper@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace Drewlabs\Contracts\Support\Actions;
 
 use Drewlabs\Contracts\Support\ArrayableInterface;
@@ -7,9 +18,9 @@ use Drewlabs\Contracts\Support\ArrayableInterface;
 interface ActionPayload extends ArrayableInterface
 {
     /**
-     * Returns the payload as an array object
-     * 
-     * @return array 
+     * Returns the payload as an array object.
+     *
+     * @return array
      */
     public function all();
 }
