@@ -1,13 +1,24 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file is part of the Drewlabs package.
+ *
+ * (c) Sidoine Azandrew <azandrewdevelopper@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace Drewlabs\Contracts\Support\Actions;
 
 interface ActionResult
 {
     /**
-     * Retuns the value holds in the action result
-     * 
-     * @return mixed 
+     * Retuns the value holds in the action result.
+     *
+     * @return mixed
      */
     public function value();
 }
