@@ -69,40 +69,42 @@ interface CollectionInterface extends \IteratorAggregate
     public function toArray(): array;
 
     /**
-     * Returns the firt item in the collection matching user specify callback if any
-     * 
-     * @param callable|\Closure|null $callback 
-     * @param mixed|null $default 
-     * @return mixed 
+     * Returns the firt item in the collection matching user specify callback if any.
+     *
+     * @param callable|\Closure|null $callback
+     * @param mixed|null             $default
+     *
+     * @return mixed
      */
     public function first($callback = null, $default = null);
 
     /**
-     * Returns the last item in the collection
-     * 
-     * @return mixed 
+     * Returns the last item in the collection.
+     *
+     * @return mixed
      */
     public function last();
 
     /**
-     * Combine the user provided keys with the current collection values
-     * 
-     * @param array|\ArrayAccess|mixed $keys 
-     * @return self 
+     * Combine the user provided keys with the current collection values.
+     *
+     * @param array|\ArrayAccess|mixed $keys
+     *
+     * @return self
      */
     public function combine($keys);
 
     /**
-     * Returns all the elements of the collection in their raw state
-     * 
-     * @return array 
+     * Returns all the elements of the collection in their raw state.
+     *
+     * @return array
      */
     public function all();
 
     /**
-     * Returns the number of elements in the collection
-     * 
-     * @return int 
+     * Returns the number of elements in the collection.
+     *
+     * @return int
      */
     public function count();
 }
