@@ -1,9 +1,19 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file is part of the Drewlabs package.
+ *
+ * (c) Sidoine Azandrew <azandrewdevelopper@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace Drewlabs\Contracts\Support\DataTransfertObject;
 
 use Drewlabs\Contracts\Support\ArrayableInterface;
-
 
 interface ObjectBridgeInterface
 {
@@ -11,7 +21,7 @@ interface ObjectBridgeInterface
      * Generate an instance of ObjectInterface.
      *
      * @param ArrayableInterface|\stdClass|array $model
-     * @param bool                  $all
+     * @param bool                               $all
      *
      * @return ObjectInterface
      */
@@ -20,17 +30,15 @@ interface ObjectBridgeInterface
     /**
      * Generate a {ArrayableInterface} from a {ObjectInterface} instance.
      *
-     * @param ObjectInterface|\stdClass|array $model
-     * 
      * @return ArrayableInterface
      */
     public function objectToModel($obj);
 
     /**
-     * Bind the output type/object/interface to the bridge class
-     * 
+     * Bind the output type/object/interface to the bridge class.
+     *
      * @param string $name
-     * 
+     *
      * @return self
      */
     public function bindClass($name);
