@@ -1,5 +1,16 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file is part of the Drewlabs package.
+ *
+ * (c) Sidoine Azandrew <azandrewdevelopper@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace Drewlabs\Contracts\Support\DataTransfertObject;
 
 use Drewlabs\Contracts\Support\ArrayableInterface;
@@ -23,15 +34,15 @@ interface ObjectInterface extends \JsonSerializable
     public function toModel();
 
     /**
-     * Convert the current object into an {@link ArrayableInterface} class
-     * 
-     * @return ArrayableInterface 
+     * Convert the current object into an {@link ArrayableInterface} class.
+     *
+     * @return ArrayableInterface
      */
     public function toArrayable();
 
     /**
      * Build the required properties of the current object from $attributes.
-     * 
+     *
      * Note: If all is true, the guarded properties are updated as well if present
      *
      * @param bool $all
