@@ -1,11 +1,20 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file is part of the Drewlabs package.
+ *
+ * (c) Sidoine Azandrew <azandrewdevelopper@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace Drewlabs\Contracts\Support\DataTransfertObject;
 
 use Drewlabs\Contracts\Support\GenericInterface;
 
-/** @package Drewlabs\Contracts\Support\DataTransfertObject */
 interface QueryResultInterface extends ObjectBridgeInterface, ListBridgeInterface, GenericInterface
 {
-
 }
