@@ -1,13 +1,22 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file is part of the Drewlabs package.
+ *
+ * (c) Sidoine Azandrew <azandrewdevelopper@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace Drewlabs\Contracts\Support;
 
 interface FuncArgument
 {
     /**
-     * Is argument optional or not
-     * 
-     * @return bool 
+     * Is argument optional or not.
      */
     public function isOptional(): bool;
 
