@@ -29,11 +29,13 @@ interface OverloadedMethodHandler
      */
     public function matches(array $args = []);
 
-     /**
-      * Call a given function or method with a list of arguments
-      
-      * @param mixed $args 
-      * @return mixed 
-      */
+    /**
+     * Call a given function or method with a list of arguments.
+
+     *
+     * @param mixed $args
+     *
+     * @return mixed
+     */
     public function call($args);
 }
