@@ -1,11 +1,20 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file is part of the Drewlabs package.
+ *
+ * (c) Sidoine Azandrew <azandrewdevelopper@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace Drewlabs\Contracts\Support;
 
 /**
- * Defines a type that reference methods with optional parameters
- * 
- * @package Drewlabs\Contracts\Support
+ * Defines a type that reference methods with optional parameters.
  */
 interface OverloadedPartialMethodHandler extends OverloadedMethodHandler
 {
