@@ -1,5 +1,16 @@
 <?php
 
+declare(strict_types=1);
+
+/*
+ * This file is part of the Drewlabs package.
+ *
+ * (c) Sidoine Azandrew <azandrewdevelopper@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace Drewlabs\Contracts\Support\Proxy;
 
 interface DynamicMethodProxyInterface
@@ -7,12 +18,13 @@ interface DynamicMethodProxyInterface
     /**
      * Provide a dynamic method proxy interface to the current object.
      * if the specified method does not exists, the default method is called
-     * instead
-     * 
-     * @param mixed $method 
-     * @param array $args 
+     * instead.
+     *
+     * @param mixed    $method
+     * @param array    $args
      * @param \Closure $default
-     * @return mixed 
+     *
+     * @return mixed
      */
-    public function call($method, $args = [], \Closure $default = null);
+    public function call($method, $args = [], ?\Closure $default = null);
 }
